Refactor Debug button into separate component
- Moved the Debug button logic from App.tsx into a new DebugButton component for better modularity and code organization.
